refactor: extract dashboard analysis helpers
