+++ scripts/check_release.py
@@ -102,6 +102,7 @@ REQUIRED_FILES = [
     "src/codex_usage_tracker/plugin_data/dashboard/dashboard_responsive.css",
     "src/codex_usage_tracker/plugin_data/dashboard/dashboard_format.js",
     "src/codex_usage_tracker/plugin_data/dashboard/dashboard_data.js",
+    "src/codex_usage_tracker/plugin_data/dashboard/dashboard_analysis.js",
     "src/codex_usage_tracker/plugin_data/dashboard/dashboard_cells.js",
     "src/codex_usage_tracker/plugin_data/dashboard/dashboard_filters.js",
     "src/codex_usage_tracker/plugin_data/dashboard/dashboard_call_investigator.js",
@@ -132,6 +133,7 @@ WHEEL_REQUIRED_MEMBERS = {
     "codex_usage_tracker/plugin_data/dashboard/dashboard_responsive.css",
     "codex_usage_tracker/plugin_data/dashboard/dashboard_format.js",
     "codex_usage_tracker/plugin_data/dashboard/dashboard_data.js",
+    "codex_usage_tracker/plugin_data/dashboard/dashboard_analysis.js",
     "codex_usage_tracker/plugin_data/dashboard/dashboard_cells.js",
     "codex_usage_tracker/plugin_data/dashboard/dashboard_filters.js",
     "codex_usage_tracker/plugin_data/dashboard/dashboard_call_investigator.js",
