+++ src/codex_usage_tracker/dashboard.py
@@ -223,6 +223,7 @@ def generate_dashboard(
     ]
     format_script_src = _versioned_asset_href(output_path, asset_base, "dashboard_format.js")
     data_script_src = _versioned_asset_href(output_path, asset_base, "dashboard_data.js")
+    analysis_script_src = _versioned_asset_href(output_path, asset_base, "dashboard_analysis.js")
     cells_script_src = _versioned_asset_href(output_path, asset_base, "dashboard_cells.js")
     filters_script_src = _versioned_asset_href(output_path, asset_base, "dashboard_filters.js")
     state_script_src = _versioned_asset_href(output_path, asset_base, "dashboard_state.js")
@@ -263,6 +264,7 @@ def generate_dashboard(
             stylesheet_hrefs=stylesheet_hrefs,
             format_script_src=format_script_src,
             data_script_src=data_script_src,
+            analysis_script_src=analysis_script_src,
             cells_script_src=cells_script_src,
             filters_script_src=filters_script_src,
             state_script_src=state_script_src,
@@ -286,6 +288,7 @@ def render_dashboard_html(
     stylesheet_hrefs: Sequence[str] | None = None,
     format_script_src: str | None = None,
     data_script_src: str | None = None,
+    analysis_script_src: str | None = None,
     cells_script_src: str | None = None,
     filters_script_src: str | None = None,
     state_script_src: str | None = None,
@@ -317,6 +320,8 @@ def render_dashboard_html(
         or _versioned_asset_href(output_path, asset_base, "dashboard_format.js"),
         data_script_src=data_script_src
         or _versioned_asset_href(output_path, asset_base, "dashboard_data.js"),
+        analysis_script_src=analysis_script_src
+        or _versioned_asset_href(output_path, asset_base, "dashboard_analysis.js"),
         cells_script_src=cells_script_src
         or _versioned_asset_href(output_path, asset_base, "dashboard_cells.js"),
         filters_script_src=filters_script_src
@@ -548,6 +553,7 @@ def _html(
     stylesheet_hrefs: Sequence[str] = ("codex-usage-tracker-assets/dashboard.css",),
     format_script_src: str = "codex-usage-tracker-assets/dashboard_format.js",
     data_script_src: str = "codex-usage-tracker-assets/dashboard_data.js",
+    analysis_script_src: str = "codex-usage-tracker-assets/dashboard_analysis.js",
     cells_script_src: str = "codex-usage-tracker-assets/dashboard_cells.js",
     filters_script_src: str = "codex-usage-tracker-assets/dashboard_filters.js",
     state_script_src: str = "codex-usage-tracker-assets/dashboard_state.js",
@@ -581,6 +587,7 @@ def _html(
         .replace("__PAYLOAD__", payload)
         .replace("__FORMAT_SCRIPT_SRC__", html.escape(format_script_src, quote=True))
         .replace("__DATA_SCRIPT_SRC__", html.escape(data_script_src, quote=True))
+        .replace("__ANALYSIS_SCRIPT_SRC__", html.escape(analysis_script_src, quote=True))
         .replace("__CELLS_SCRIPT_SRC__", html.escape(cells_script_src, quote=True))
         .replace("__FILTERS_SCRIPT_SRC__", html.escape(filters_script_src, quote=True))
         .replace("__STATE_SCRIPT_SRC__", html.escape(state_script_src, quote=True))
