+++ src/codex_usage_tracker/plugin_data/dashboard/dashboard_template.html
@@ -155,6 +155,7 @@
   <script id="usage-data" type="application/json">__PAYLOAD__</script>
   <script src="__FORMAT_SCRIPT_SRC__"></script>
   <script src="__DATA_SCRIPT_SRC__"></script>
+  <script src="__ANALYSIS_SCRIPT_SRC__"></script>
   <script src="__CELLS_SCRIPT_SRC__"></script>
   <script src="__FILTERS_SCRIPT_SRC__"></script>
   <script src="__STATE_SCRIPT_SRC__"></script>
